corrections on like functions

+++ src/main/java/com/project/socialmediaplatform/controller/PostController.java
@@ -26,7 +26,7 @@ import com.project.socialmediaplatform.service.PostService;
 
 @RestController
 @RequestMapping("/api/user")
-public class PostController extends UserManager{
+public class PostController extends UserManager {
 
     @Autowired
     private PostService postService;
@@ -38,8 +38,8 @@ public class PostController extends UserManager{
     @PostMapping("/post")
     public ResponseEntity<Post> createPost(@RequestBody Post newPost) {
         User user = getUserFromAuthentication();
-        if(user==null)
-        throw new UserNotFoundException("no user");
+        if (user == null)
+            throw new UserNotFoundException("no user");
         if (newPost.getContent() == null) {
             throw new ValidationException("Post is Empty");
         }
@@ -56,11 +56,16 @@ public class PostController extends UserManager{
     public ResponseEntity<String> deletePost(@PathVariable Long postId) {
         User user = getUserFromAuthentication();
         Post post = postRepo.findByPostId(postId);
-        if(post.getUser()==user){
-        postService.deletePost(postId);
-        return ResponseEntity.ok("Deleted Successfully");
+        if (post != null) {
+            if (post.getUser() == user) {
+                postService.deletePost(postId);
+                return ResponseEntity.ok("Deleted Successfully");
+            } 
+            else
+                throw new AuthorizationServiceException("Unauthorized action");
         }
-        else throw new AuthorizationServiceException("Unauthorized action");
+        else
+            throw new PostNotFoundException("No post found");
     }
 
     // editpost
@@ -68,29 +73,31 @@ public class PostController extends UserManager{
     public ResponseEntity<Post> editPost(@PathVariable Long postId, @RequestBody Post postUpdate) {
         User user = getUserFromAuthentication();
         Post post = postRepo.findByPostId(postId);
-        if(post==null)
-        throw new PostNotFoundException("No such post Exists");
-        if(post.getUser()==user){
+        if (post == null)
+            throw new PostNotFoundException("No such post Exists");
+        if (post.getUser() == user) {
             Post editedPost = postService.editPost(post, postUpdate.getCaption());
             return ResponseEntity.ok(editedPost);
-        }
-        else throw new AuthorizationServiceException("Unauthorized action");
+        } else
+            throw new AuthorizationServiceException("Unauthorized action");
     }
 
     // // viewposts
     // @GetMapping("/{userId}/post")
-    // public ResponseEntity<List<Post>> getOtherUserPosts(@PathVariable Long userId) {
-    //     List<Post> otherUserPosts = userService.getOtherUserPosts(userId);
-    //     return ResponseEntity.ok(otherUserPosts);
+    // public ResponseEntity<List<Post>> getOtherUserPosts(@PathVariable Long
+    // userId) {
+    // List<Post> otherUserPosts = userService.getOtherUserPosts(userId);
+    // return ResponseEntity.ok(otherUserPosts);
     // }
 
     @GetMapping("/post/search")
-    public ResponseEntity<List<Post>> searchPost(@ModelAttribute("search") SearchModel searchModel) { 
+    public ResponseEntity<List<Post>> searchPost(@ModelAttribute("search") SearchModel searchModel) {
         User user = getUserFromAuthentication();
-        if(searchModel.getSearchWord()==null && searchModel.getUserId()==null && searchModel.getPostDate()==null){
+        if (searchModel.getSearchWord() == null && searchModel.getUserId() == null
+                && searchModel.getPostDate() == null) {
             return null;
-        }  
-        List<Post> searchResults = postService.searchPosts(user,searchModel);
+        }
+        List<Post> searchResults = postService.searchPosts(user, searchModel);
         return ResponseEntity.ok(searchResults);
     }
